Move deprecated-proposed arbitraries tests to `shared` tests file
